PythonWebApp/app.py
from flask import Flask, request, jsonify
from flask_cors import CORS
import google.generativeai as genai

app = Flask(__name__)
CORS(app)

# Configure Gemini API - Add your API key here
GEMINI_API_KEY = ""
if GEMINI_API_KEY:
    genai.configure(api_key=GEMINI_API_KEY)


# Client function that calls Google Gemini API
def client_function(user_input: str) -> str:
    """
    Sends user input to Google Gemini API and returns the response.
    """
    if not GEMINI_API_KEY:
        raise ValueError("GEMINI_API_KEY is not configured. Please add your API key.")
    
    try:
        model = genai.GenerativeModel('gemini-2.5-flash')
        response = model.generate_content(user_input)
        return response.text
    except Exception as e:
        raise Exception(f"Gemini API error: {str(e)}")


# Stubbed out API client function
def api_client_function(processed_data: str) -> str:
    """
    Takes processed data and makes an API call.
    This is a stub that you can replace with actual API logic.
    """
    # Placeholder implementation
    return f"API Result: {processed_data} - [API Response]"


@app.route('/api/process', methods=['POST'])
def process_input():
    """
    Endpoint to process user input through the client function.
    """
    try:
        data = request.get_json()
        user_input = data.get('input', '')
        
        if not user_input:
            return jsonify({'error': 'Input is required'}), 400
        
        result = client_function(user_input)
        return jsonify({'result': result}), 200
    except Exception as e:
        return jsonify({'error': str(e)}), 500


@app.route('/api/call', methods=['POST'])
def call_api():
    """
    Endpoint to trigger the API client function with processed data.
    """
    try:
        data = request.get_json()
        processed_data = data.get('data', '')
        
        if not processed_data:
            return jsonify({'error': 'Data is required'}), 400
        
        result = api_client_function(processed_data)
        return jsonify({'result': result}), 200
    except Exception as e:
        return jsonify({'error': str(e)}), 500


if __name__ == '__main__':
    app.run(debug=True, port=5000)
